backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic[email]==2.4.2
scikit-learn==1.3.1
pandas==2.1.1
numpy==1.26.0
joblib==1.3.2
python-multipart==0.0.6
sqlalchemy==2.0.22
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-dotenv==1.0.0
httpx==0.24.0
pytest==7.3.1
matplotlib==3.7.2
